chore: remove schemaUpdater in favor of MS
