+++ package.json
@@ -7,7 +7,6 @@
     "lint": "eslint --fix . && prettier --write **/*.ts **/{metadata,tsconfig}.json",
     "lint:ci": "eslint . && prettier **/*.ts **/{metadata,tsconfig}.json",
     "inc": "tsc -p ./util/configs/tsconfig.massVer.json && node ./util/tools/massVer.js",
-    "su": "tsc -p ./util/configs/tsconfig.schemaUpdater.json && node ./util/tools/schemaUpdater.js",
     "se": "tsc -p ./util/configs/tsconfig.auto.syntaxEnforcer.json && node ./util/tools/auto/syntaxEnforcer.js",
     "sv": "tsc -p ./util/configs/tsconfig.auto.schemaValidator.json && node ./util/tools/auto/schemaValidator.js",
     "pu": "tsc -p ./util/configs/tsconfig.auto.presenceUpdater.json && node ./util/tools/auto/presenceUpdater.js",
