--- util/configs/tsconfig.schemaUpdater.json
@@ -1,4 +1,0 @@
-{
-  "extends": "../tools/auto/tsconfig.auto.json",
-  "files": ["../tools/schemaUpdater.ts"]
-}
